feat!: Parse integer variables as int64 and float variables as float64

+++ gqt_test.go
@@ -30,7 +30,7 @@ var tests = []ExpectDoc{
 						InputConstraints: []gqt.InputConstraint{{
 							Name: "x1",
 							Constraint: gqt.ConstraintValEqual{
-								Value: float64(1),
+								Value: int64(1),
 							},
 						}},
 						Selections: []gqt.Selection{
@@ -148,7 +148,7 @@ var tests = []ExpectDoc{
 							{
 								Name: "x1",
 								Constraint: gqt.ConstraintValEqual{
-									Value: float64(1),
+									Value: int64(1),
 								},
 							},
 						},
@@ -168,12 +168,12 @@ var tests = []ExpectDoc{
 	Expect(`mutation {
 		a(
 			y1: any
-			i1: val = 42
-			i2: val != 42
+			i1: val = 42.0
+			i2: val != 42.0
 			i3: val > 42
-			i4: val < 42
+			i4: val < 42.0
 			i5: val >= 42
-			i6: val <= 42
+			i6: val <= 42.0
 			i7: val = "text"
 			i8: val != "text"
 			l1: len = 42
@@ -189,7 +189,7 @@ var tests = []ExpectDoc{
 			b5: bytelen >= 42
 			b6: bytelen <= 42
 			a1: val = []
-			a2: val = [ ... val < 10 ]
+			a2: val = [ ... val < 10.0 ]
 			a3: val = [ val = "a", val = "b" ]
 			o1: val = {
 				of1: val = true
@@ -220,7 +220,7 @@ var tests = []ExpectDoc{
 				}, {
 					Name: "i3",
 					Constraint: gqt.ConstraintValGreater{
-						Value: float64(42),
+						Value: int64(42),
 					},
 				}, {
 					Name: "i4",
@@ -230,7 +230,7 @@ var tests = []ExpectDoc{
 				}, {
 					Name: "i5",
 					Constraint: gqt.ConstraintValGreaterOrEqual{
-						Value: float64(42),
+						Value: int64(42),
 					},
 				}, {
 					Name: "i6",
@@ -316,7 +316,7 @@ var tests = []ExpectDoc{
 					Name: "a2",
 					Constraint: gqt.ConstraintMap{
 						Constraint: gqt.ConstraintValLess{
-							Value: 10,
+							Value: 10.0,
 						},
 					},
 				}, {
@@ -370,13 +370,13 @@ var tests = []ExpectDoc{
 	}),
 	Expect(`mutation {
 		a(
-			a: val > 0 && val < 3
-			b: val > 0 && val < 9 && val != 5
-			c: val = 1 || val = 2
-			d: val = 1 || val = 2 || val = 3
+			a: val > 0.0 && val < 3.0
+			b: val > 0.0 && val < 9.0 && val != 5.0
+			c: val = 1.0 || val = 2.0
+			d: val = 1.0 || val = 2.0 || val = 3.0
 			e: bytelen > 3 && bytelen <= 10 ||
 				val = true ||
-				val = 1
+				val = 1.0
 		) {b}
 	}`, gqt.DocMutation{
 		Selections: []gqt.Selection{
